Bluetooth control implemented
