+++ src/fx.c
@@ -1,5 +1,7 @@
 #include "fx.h"
 
+#include "interface/rgb.h"
+
 #include "string.h"
 
 uint16_t fx_select_effect(effect index, uint16_t signal) {
@@ -42,4 +44,20 @@ effect fx_from_string(const char* str) {
     }
 
     return UNKNOWN;
+}
+
+effect_rgb fx_rgb_color(effect index) {
+    switch (index) {
+        case REVERB: return (effect_rgb) {MAX_RGB_COLOR, 0, 0};
+        case DELAY: return (effect_rgb) {0, MAX_RGB_COLOR, 0};
+        case ECHO: return (effect_rgb) {0, 0, MAX_RGB_COLOR};
+        case BOOSTER: return (effect_rgb) {MAX_RGB_COLOR, MAX_RGB_COLOR, 0};
+        case FUZZ: return (effect_rgb) {0, MAX_RGB_COLOR, MAX_RGB_COLOR};
+        case DISTORTION: return (effect_rgb) {MAX_RGB_COLOR, 0, MAX_RGB_COLOR};
+        case BITCRUSH: return (effect_rgb) {MAX_RGB_COLOR/2, MAX_RGB_COLOR/4, 0};
+        case TREMOLO: return (effect_rgb) {MAX_RGB_COLOR/3, 0, MAX_RGB_COLOR/3};
+        case OCTAVER: return (effect_rgb) {MAX_RGB_COLOR/3, MAX_RGB_COLOR/2, MAX_RGB_COLOR/5};
+    }
+
+    return (effect_rgb) {MAX_RGB_COLOR, MAX_RGB_COLOR, MAX_RGB_COLOR};
 }
