+++ src/interface/i2c.h
@@ -1,0 +1,8 @@
+#ifndef SOUNDPI_PFX_I2C_H
+#define SOUNDPI_PFX_I2C_H
+
+#include "conf.h"
+
+void init_i2c_communication(void);
+
+#endif //SOUNDPI_PFX_I2C_H
