+++ src/main.c
@@ -5,34 +5,22 @@
 #include <pico/stdlib.h>
 #include <pico/multicore.h>
 
-#include "hardware/pwm.h"
-#include "hardware/uart.h"
-#include "hardware/i2c.h"
-
 #include "conf.h"
 
 #include "interface/audio.h"
-#include "interface/screen.h"
 #include "interface/bluetooth.h"
 #include "interface/rgb.h"
 #include "interface/blink.h"
 
-#include "low/pwm.h"
-
 #include "fx.h"
-#include "umath.h"
 #include "ustr.h"
-
-#define I2C_PORT        i2c0
-#define I2C_ADDRESS     0x3C
 
 #define BUFF_INPUT_LENGHT       86
 
 static uint16_t signal;
 
-void init_i2c_communication(void);
-
 effect current_effect = CLEAN;
+effect_rgb current_color = {MAX_RGB_COLOR, MAX_RGB_COLOR, MAX_RGB_COLOR};
 
 static void core1(void) {
     init_blink();
@@ -47,8 +35,8 @@ static void core1(void) {
             effect effect_selected = fx_from_string(buffinput);
 
             if(effect_selected != UNKNOWN) {
-
                 current_effect = fx_from_string(buffinput);
+                current_color = fx_rgb_color(current_effect);
 
                 ble_send_str("OK+Effect=");
                 ble_send_str(buffinput);
@@ -61,6 +49,8 @@ static void core1(void) {
 
             memset(buffinput, 0, BUFF_INPUT_LENGHT);
         }
+
+        rgb_set_color(current_color.r, current_color.g, current_color.b);
 
         blink();
     }
@@ -90,14 +80,4 @@ int main() {
 
         write_audio(signal);
     }
-}
-
-void init_i2c_communication(void) {
-    i2c_init(I2C_PORT, 100 * 1000);
-
-    gpio_set_function(4, GPIO_FUNC_I2C);
-    gpio_set_function(5, GPIO_FUNC_I2C);
-
-    gpio_pull_up(4);
-    gpio_pull_up(5);
 }
